Join more CaseChildElement/LedgerChildElement cod into StorageBackedChildElement
